custom hook to retrieve and update tab state
